[refactor] Divide interface for Service 서비스 분리 related to #36
