+++ search/src/main/java/com/bjcareer/search/application/exceptions/InvalidStockInformation.java
@@ -1,4 +1,4 @@
-package com.bjcareer.search.service.exceptions;
+package com.bjcareer.search.application.exceptions;
 
 public class InvalidStockInformation extends StockServiceException {
 	public InvalidStockInformation(String message) {
