+++ search/src/main/java/com/bjcareer/search/application/port/in/AddStockUsecase.java
@@ -1,0 +1,7 @@
+package com.bjcareer.search.application.port.in;
+
+import com.bjcareer.search.domain.entity.Thema;
+
+public interface AddStockUsecase {
+	Thema addStockThema(String stockCode, String stockName, String themeName);
+}
