+++ search/src/main/java/com/bjcareer/search/application/port/in/InformationUsecase.java
@@ -1,0 +1,10 @@
+package com.bjcareer.search.application.port.in;
+
+import java.time.LocalDate;
+import java.util.Map;
+
+import com.bjcareer.search.domain.GTPNewsDomain;
+
+public interface InformationUsecase {
+	Map<LocalDate, GTPNewsDomain> findSearchRaiseReason(String stockName);
+}
