+++ search/src/main/java/com/bjcareer/search/application/port/in/KeywordUsecase.java
@@ -1,0 +1,9 @@
+package com.bjcareer.search.application.port.in;
+
+import java.util.List;
+
+import com.bjcareer.search.domain.AbsoluteRankKeyword;
+
+public interface KeywordUsecase {
+	List<AbsoluteRankKeyword> getAbsoluteValueOfKeyword(String keyword);
+}
