+++ search/src/main/java/com/bjcareer/search/application/port/in/ReasonUsecase.java
@@ -1,0 +1,10 @@
+package com.bjcareer.search.application.port.in;
+
+import java.time.LocalDate;
+import java.util.Map;
+
+import com.bjcareer.search.domain.GTPNewsDomain;
+
+public interface ReasonUsecase {
+	Map<LocalDate, GTPNewsDomain> findSearchRaiseReason(String stockName);
+}
