+++ search/src/main/java/com/bjcareer/search/application/stock/StockService.java
@@ -1,10 +1,11 @@
-package com.bjcareer.search.service;
+package com.bjcareer.search.application.stock;
 
 import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.bjcareer.search.application.port.in.AddStockUsecase;
 import com.bjcareer.search.domain.entity.Stock;
 import com.bjcareer.search.domain.entity.Thema;
 import com.bjcareer.search.domain.entity.ThemaInfo;
@@ -12,7 +13,7 @@ import com.bjcareer.search.out.crawling.naver.CrawlingNaverFinance;
 import com.bjcareer.search.repository.stock.StockRepository;
 import com.bjcareer.search.repository.stock.ThemaInfoRepository;
 import com.bjcareer.search.repository.stock.ThemaRepository;
-import com.bjcareer.search.service.exceptions.InvalidStockInformation;
+import com.bjcareer.search.application.exceptions.InvalidStockInformation;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -21,7 +22,7 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 @Slf4j
 @Transactional
-public class StockService {
+public class StockService implements AddStockUsecase {
 	private final ThemaRepository themaRepository;
 	private final ThemaInfoRepository themaInfoRepository;
 	private final StockRepository stockRepository;
